backend/pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "marconi-elearn-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic-settings>=2.2",
  "email-validator>=2.1",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.1",
  "asyncpg>=0.29",
  "python-multipart>=0.0.9",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "cryptography>=42.0",
  "taskiq>=0.12.1",
  "taskiq-redis>=1.2.1",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4",
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]

[tool.ruff]
line-length = 100

[tool.setuptools]
packages = ["app"]
